# platform/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "DesignPromptIn": {
        "properties": {
          "prompt": {
            "maxLength": 30000,
            "minLength": 20,
            "title": "Prompt",
            "type": "string"
          }
        },
        "required": [
          "prompt"
        ],
        "title": "DesignPromptIn",
        "type": "object"
      },
      "DesignPromptOut": {
        "properties": {
          "edited_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edited At"
          },
          "edited_by_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edited By Name"
          },
          "organization_name": {
            "title": "Organization Name",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "saved": {
            "title": "Saved",
            "type": "boolean"
          }
        },
        "required": [
          "project_id",
          "organization_name",
          "prompt",
          "saved",
          "edited_at",
          "edited_by_name"
        ],
        "title": "DesignPromptOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PageInfo": {
        "properties": {
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "next_cursor",
          "has_more",
          "limit"
        ],
        "title": "PageInfo",
        "type": "object"
      },
...
      "ProjectOut": {
        "properties": {
          "available_actions": {
            "items": {
              "type": "string"
            },
            "title": "Available Actions",
            "type": "array"
          },
          "brief": {
            "additionalProperties": true,
            "title": "Brief",
            "type": "object"
          },
          "brief_submitted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brief Submitted At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "due_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due On"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "live_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Live Url"
          },
          "missing_fields": {
            "items": {
              "type": "string"
            },
            "title": "Missing Fields",
            "type": "array"
          },
          "organization_id": {
            "format": "uuid",
            "title": "Organization Id",
            "type": "string"
          },
          "organization_name": {
            "title": "Organization Name",
            "type": "string"
          },
          "preview_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview Url"
          },
          "revision_requests": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Revision Requests",
            "type": "array"
          },
          "revisions_left": {
            "title": "Revisions Left",
            "type": "integer"
          },
          "revisions_used": {
            "title": "Revisions Used",
            "type": "integer"
          },
          "scope_accepted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope Accepted At"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "organization_id",
          "organization_name",
          "status",
          "brief",
          "missing_fields",
          "scope_accepted_at",
          "brief_submitted_at",
          "due_on",
          "preview_url",
          "live_url",
          "revisions_used",
          "revisions_left",
          "revision_requests",
          "available_actions",
          "created_at",
          "updated_at"
        ],
        "title": "ProjectOut",
        "type": "object"
      },
      "ProjectPage": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ProjectOut"
            },
            "title": "Data",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "required": [
          "data",
          "page"
        ],
        "title": "ProjectPage",
        "type": "object"
      },
      "ProjectStatus": {
        "enum": [
          "BRIEF_PENDING",
          "IN_PRODUCTION",
          "CLIENT_REVIEW",
          "REVISION",
          "APPROVED",
          "LIVE",
          "CANCELLED"
        ],
        "title": "ProjectStatus",
        "type": "string"
      },
...
      "RevisionRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RevisionItem"
            },
            "maxItems": 10,
            "minItems": 1,
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "RevisionRequest",
        "type": "object"
      },
...
      "SubmitRequest": {
        "properties": {
          "accept_scope": {
            "title": "Accept Scope",
            "type": "boolean"
          }
        },
        "required": [
          "accept_scope"
        ],
        "title": "SubmitRequest",
        "type": "object"
      },
...
      "TransitionRequest": {
        "properties": {
          "live_url": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 2083,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Live Url"
          },
          "preview_url": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 2083,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview Url"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "target": {
            "$ref": "#/components/schemas/ProjectStatus"
          }
        },
        "required": [
          "target"
        ],
        "title": "TransitionRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "BENILAB Digital360 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/website-projects": {
      "get": {
        "operationId": "admin_list_api_v1_admin_website_projects_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ProjectStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin List",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/website-projects/{project_id}": {
      "get": {
        "operationId": "admin_get_api_v1_admin_website_projects__project_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Get",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/website-projects/{project_id}/design-prompt": {
      "get": {
        "description": "Prompt de conception du site (texte à améliorer avant de lancer l'outil d'IA).",
        "operationId": "admin_get_design_prompt_api_v1_admin_website_projects__project_id__design_prompt_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DesignPromptOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Get Design Prompt",
        "tags": [
          "admin"
        ]
      },
      "put": {
        "operationId": "admin_save_design_prompt_api_v1_admin_website_projects__project_id__design_prompt_put",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DesignPromptIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DesignPromptOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Save Design Prompt",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/website-projects/{project_id}/design-prompt.pdf": {
      "get": {
        "operationId": "admin_design_prompt_pdf_api_v1_admin_website_projects__project_id__design_prompt_pdf_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/pdf": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Design Prompt Pdf",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/website-projects/{project_id}/design-prompt/regenerate": {
      "post": {
        "description": "Réécrit le prompt depuis le brief actuel (remplace les modifications de l'équipe).",
        "operationId": "admin_regenerate_design_prompt_api_v1_admin_website_projects__project_id__design_prompt_regenerate_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DesignPromptOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Regenerate Design Prompt",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/website-projects/{project_id}/transition": {
      "post": {
        "operationId": "admin_transition_api_v1_admin_website_projects__project_id__transition_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Transition",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/orgs/{org_id}/website-project": {
      "get": {
        "operationId": "get_project_api_v1_orgs__org_id__website_project_get",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientProjectOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Project",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/orgs/{org_id}/website-project/approve": {
      "post": {
        "operationId": "approve_api_v1_orgs__org_id__website_project_approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientProjectOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/orgs/{org_id}/website-project/brief": {
      "put": {
        "description": "Enregistre le brouillon (partiel accepté, mais toujours dans le cadre de l'offre).",
        "operationId": "save_brief_api_v1_orgs__org_id__website_project_brief_put",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Brief"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientProjectOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Brief",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/orgs/{org_id}/website-project/revision": {
      "post": {
        "operationId": "request_revision_api_v1_orgs__org_id__website_project_revision_post",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientProjectOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request Revision",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/orgs/{org_id}/website-project/submit": {
      "post": {
        "description": "Brief complet et périmètre accepté : la production démarre, la date de livraison est fixée.",
        "operationId": "submit_brief_api_v1_orgs__org_id__website_project_submit_post",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientProjectOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Brief",
        "tags": [
          "projects"
        ]
      }
    },
...
  }
}